调整税率浮点数和读取玩家限制下注数据

+++ game/internal/configData.go
@@ -41,7 +41,7 @@ func getNextTime() string {
 	return datetime + ":00"
 }
 
-func SetPackageTaxM(packageT uint16, tax uint8) {
+func SetPackageTaxM(packageT uint16, tax float64) {
 	packageTax[packageT] = tax
 }
